# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText
  description: |-
    This OpenAPI document describes APIs for the project "WASAText".

    Student: Aleksandar Kirilov
    Matricola: 2059920
    E-mail: kirilov.2059920@studenti.uniroma1.it
  version: "1.0.2"
...
security:
- BearerAuth: []
paths:
...
  /conversations:
    get:
      tags:
      - conversation
      summary: Retrieves the list of conversations for the authenticated user
      description: Retrieves conversations.
      operationId: getMyConversations
      security:
      - BearerAuth: []
      responses:
        '200':
          description: List of conversations
          content:
            application/json:
              schema:
                type: object
                description: Contains a list of conversations.
                required:
                - conversations
                properties:
                  conversations:
                    type: array
                    description: List of conversations
                    minItems: 0
                    maxItems: 1000
                    items:
                      $ref: '#/components/schemas/ConversationDetailsSummary'
              example:
                conversations:
                - id: "conv123"
                  name: "Group Chat"
                  members:
                  - "user123"
                  - "user456"
                  conversationPhoto: "aGVsbG8="
                  lastMessage:
                    id: "msg789"
                    senderId: "user456"
                    senderName: "Alice"
                    content: "Hello!"
                    timestamp: "2023-10-20T10:00:00Z"
                    attachment: ""
                    reactionCount: 0
                    reactingUserIds: []
    post:
      tags:
      - conversation
      summary: Starts a new direct conversation
      description: Creates a new direct conversation between two users if one does not exist and returns the canonical conversation details.
      operationId: startConversation
      security:
      - BearerAuth: []
      requestBody:
        description: JSON payload with sender and recipient IDs.
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Payload to start a direct conversation.
              required:
              - senderId
              - recipientId
              properties:
                senderId:
                  type: string
                  description: ID of the sender.
                  example: "user123"
                  pattern: '^[a-zA-Z0-9_]+$'
                  minLength: 1
                  maxLength: 50
                recipientId:
                  type: string
                  description: ID of the recipient.
                  example: "user456"
                  pattern: '^[a-zA-Z0-9_]+$'
                  minLength: 1
                  maxLength: 50
      responses:
        '201':
          description: Conversation started successfully and returns the canonical conversation details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationDetails'
              example:
                id: "conv123"
                name: "Group Chat"
                members:
                - "user123"
                - "user456"
                conversationPhoto: "aGVsbG8="
                lastMessage:
                  id: "msg789"
                  senderId: "user456"
                  senderName: "Alice"
                  content: "Hello!"
                  timestamp: "2023-10-20T10:00:00Z"
                  attachment: ""
                  reactionCount: 0
                  reactingUserIds: []
                messages: []
  /conversations/{conversationId}:
    get:
      tags:
      - conversation
      summary: Retrieves details of a specific conversation
      description: Retrieves conversation details.
      operationId: getConversation
      security:
      - BearerAuth: []
      parameters:
      - name: conversationId
        in: path
        required: true
        description: ID of the conversation.
        schema:
          type: string
          pattern: '^[a-zA-Z0-9_]+$'
          minLength: 1
          maxLength: 50
      responses:
        '200':
          description: Conversation details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationDetails'
              example:
                id: "conv123"
                name: "Group Chat"
                members:
                - "user123"
                - "user456"
                conversationPhoto: "aGVsbG8="
                lastMessage:
                  id: "msg789"
                  senderId: "user456"
                  senderName: "Alice"
                  content: "Hello!"
                  timestamp: "2023-10-20T10:00:00Z"
                  attachment: ""
                  reactionCount: 0
                  reactingUserIds: []
                messages: []
  /conversations/{conversationId}/message:
    post:
      tags:
      - message
      summary: Sends a message in a conversation
      description: Sends a message using multipart/form-data.
      operationId: sendMessage
      security:
      - BearerAuth: []
      parameters:
      - name: conversationId
        in: path
        required: true
        description: ID of the conversation.
        schema:
          type: string
          pattern: '^[a-zA-Z0-9_]+$'
          minLength: 1
          maxLength: 50
      requestBody:
        description: Form data containing message content and an optional attachment.
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              description: Form data for sending a message.
              properties:
                content:
                  type: string
                  description: The content of the message.
                  example: "Hello, world!"
                  pattern: '^.*$'
                  minLength: 1
                  maxLength: 1000
                attachment:
                  type: string
                  description: Optional image attachment (JPEG, PNG, or GIF), base64 encoded.
                  example: ""
                  pattern: '^[A-Za-z0-9+/]*={0,2}$'
                  minLength: 0
                  maxLength: 10485760
      responses:
        '201':
          description: Message sent successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                id: "msg123"
                senderId: "user123"
                senderName: "Maria"
                content: "Hello, world!"
                timestamp: "2023-10-20T10:05:00Z"
                attachment: ""
                reactionCount: 0
                reactingUserIds: []
  /conversations/{conversationId}/message/{messageId}/forward:
    post:
      tags:
      - message
      summary: Forwards an existing message to another conversation
      description: Forwards a message to another conversation.
      operationId: forwardMessage
      security:
      - BearerAuth: []
      parameters:
      - name: conversationId
        in: path
        required: true
        description: ID of the conversation containing the message to forward.
        schema:
          type: string
          pattern: '^[a-zA-Z0-9_]+$'
          minLength: 1
          maxLength: 50
      - name: messageId
        in: path
        required: true
        description: ID of the message to forward.
        schema:
          type: string
          pattern: '^[a-zA-Z0-9_]+$'
          minLength: 1
          maxLength: 50
      requestBody:
        description: JSON payload specifying the target conversation ID.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForwardMessageRequest'
            example:
              targetConversationId: "conv456"
      responses:
        '201':
          description: Message forwarded successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                id: "msg999"
                senderId: "user123"
                senderName: "Maria"
                content: "Hello, world!"
                timestamp: "2023-10-20T10:05:00Z"
                attachment: ""
                reactionCount: 0
                reactingUserIds: []
  /conversations/{conversationId}/message/{messageId}:
    delete:
      tags:
      - message
      summary: Deletes a message
      description: Deletes a message from the conversation.
      operationId: deleteMessage
      security:
      - BearerAuth: []
      parameters:
      - name: conversationId
        in: path
        required: true
        description: ID of the conversation.
        schema:
          type: string
          pattern: '^[a-zA-Z0-9_]+$'
          minLength: 1
          maxLength: 50
      - name: messageId
        in: path
        required: true
        description: ID of the message to delete.
        schema:
          type: string
          pattern: '^[a-zA-Z0-9_]+$'
          minLength: 1
          maxLength: 50
      responses:
        '204':
          description: Message deleted successfully.
  /conversations/{conversationId}/message/{messageId}/comment:
    parameters:
    - name: conversationId
      in: path
      required: true
      description: ID of the conversation.
      schema:
        type: string
        pattern: '^[a-zA-Z0-9_]+$'
        minLength: 1
        maxLength: 50
    - name: messageId
      in: path
      required: true
      description: ID of the message to comment on.
      schema:
        type: string
        pattern: '^[a-zA-Z0-9_]+$'
        minLength: 1
        maxLength: 50
    post:
      tags:
      - comment
      summary: Adds a comment to a message
      description: Adds a comment to the specified message.
      operationId: commentMessage
      security:
      - BearerAuth: []
      requestBody:
        description: JSON payload with comment content.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCommentRequest'
      responses:
        '204':
          description: Comment added successfully.
    delete:
      tags:
      - comment
      summary: Removes a comment from a message
      description: Deletes a comment from the specified message.
      operationId: uncommentMessage
      security:
      - BearerAuth: []
      responses:
        '204':
          description: Comment deleted successfully.
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ConversationDetailsSummary:
      type: object
      description: A summary of a conversation between users.
      required:
      - id
      - name
      - members
      properties:
        id:
          type: string
          description: Unique identifier of the conversation.
          example: "conv123"
          pattern: '^[a-zA-Z0-9_]+$'
          minLength: 1
          maxLength: 50
        name:
          type: string
          description: Name of the conversation.
          example: "Group Chat"
          pattern: '^[a-zA-Z0-9_ ]+$'
          minLength: 1
          maxLength: 50
        members:
          type: array
          description: List of user IDs participating in the conversation.
          minItems: 1
          maxItems: 1000
          items:
            type: string
            description: User ID.
            example: "user123"
            pattern: '^[a-zA-Z0-9_]+$'
            minLength: 1
            maxLength: 50
        conversationPhoto:
          type: string
          description: Base64-encoded conversation photo (if any).
          example: "aGVsbG8="
          pattern: '^[A-Za-z0-9+/]*={0,2}$'
          minLength: 0
          maxLength: 1000000
        lastMessage:
          $ref: '#/components/schemas/Message'
    ConversationDetails:
      title: "Conversation Details"
      type: object
      description: Detailed conversation schema.
      required:
      - id
      - name
      - members
      - messages
      properties:
        id:
          type: string
          description: Unique identifier of the conversation.
          example: "conv123"
          pattern: '^[a-zA-Z0-9_]+$'
          minLength: 1
          maxLength: 50
        name:
          type: string
          description: Name of the conversation.
          example: "Group Chat"
          pattern: '^[a-zA-Z0-9_ ]+$'
          minLength: 1
          maxLength: 50
        members:
          type: array
          description: List of user IDs participating in the conversation.
          minItems: 1
          maxItems: 1000
          items:
            type: string
            description: User ID.
            example: "user123"
            pattern: '^[a-zA-Z0-9_]+$'
            minLength: 1
            maxLength: 50
        conversationPhoto:
          type: string
          description: Base64-encoded conversation photo (if any).
          example: "aGVsbG8="
          pattern: '^[A-Za-z0-9+/]*={0,2}$'
          minLength: 0
          maxLength: 1000000
        lastMessage:
          $ref: '#/components/schemas/Message'
        messages:
          type: array
          description: List of messages in the conversation.
          minItems: 0
          maxItems: 1000
          items:
            $ref: '#/components/schemas/Message'
    Message:
      type: object
      description: A message sent in a conversation.
      required:
      - id
      - senderId
      - senderName
      - content
      - timestamp
      - attachment
      - reactionCount
      - reactingUserIds
      properties:
        id:
          type: string
          description: Unique identifier of the message.
          example: "msg123"
          pattern: '^[a-zA-Z0-9_]+$'
          minLength: 1
          maxLength: 50
        senderId:
          type: string
          description: ID of the user who sent the message.
          example: "user456"
          pattern: '^[a-zA-Z0-9_]+$'
          minLength: 1
          maxLength: 50
        senderName:
          type: string
          description: Name of the sender.
          example: "Alice"
          pattern: '^[a-zA-Z0-9 ]+$'
          minLength: 1
          maxLength: 50
        content:
          type: string
          description: Content of the message.
          example: "Hello, world!"
          pattern: '^.*$'
          minLength: 1
          maxLength: 1000
        timestamp:
          type: string
          format: date-time
          description: When the message was sent.
          example: "2023-10-20T10:05:00Z"
          minLength: 20
          maxLength: 29
        attachment:
          type: string
          description: Base64-encoded attachment (if any).
          example: ""
          pattern: '^[A-Za-z0-9+/]*={0,2}$'
          minLength: 0
          maxLength: 10485760
        reactionCount:
          type: integer
          description: Number of reactions on the message.
          example: 2
        reactingUserIds:
          type: array
          description: List of user IDs who have reacted to the message.
          minItems: 0
          maxItems: 1000
          items:
            type: string
            example: "user123"
            pattern: '^[a-zA-Z0-9_]+$'
            minLength: 1
            maxLength: 50
        replyTo:
          type: string
          description: (Optional) ID of the message being replied to.
          example: ""
          pattern: '^[a-zA-Z0-9_]*$'
          minLength: 0
          maxLength: 50
        replyContent:
          type: string
          description: (Optional) A preview of the message being replied to.
          example: ""
          pattern: '^.*$'
          minLength: 0
          maxLength: 1000
        replySenderName:
          type: string
          description: (Optional) Name of the sender of the replied-to message.
          example: ""
          pattern: '^[a-zA-Z0-9 ]*$'
          minLength: 0
          maxLength: 50
        replyAttachment:
          type: string
          description: (Optional) Base64-encoded attachment from the replied-to message.
          example: ""
          pattern: '^[A-Za-z0-9+/]*={0,2}$'
          minLength: 0
          maxLength: 10485760
        status:
          type: string
          description: (Optional) Delivery/read status; provided only for messages not sent by the authenticated user.
          example: "✓"
          pattern: '^.*$'
          minLength: 0
          maxLength: 10
    ForwardMessageRequest:
      type: object
      description: Request schema for forwarding a message.
      required:
      - targetConversationId
      properties:
        targetConversationId:
          type: string
          description: ID of the conversation to which the message will be forwarded.
          example: "conv456"
          pattern: '^[a-zA-Z0-9_]+$'
          minLength: 1
          maxLength: 50
        forwarderName:
          type: string
          description: (Optional) Name of the user forwarding the message.
          example: "Maria"
          pattern: '^[a-zA-Z0-9 ]*$'
          minLength: 0
          maxLength: 50
    AddCommentRequest:
      type: object
      description: Request schema for adding a comment.
      required:
      - content
      properties:
        content:
          type: string
          description: The content of the comment.
          example: "Nice post!"
          pattern: '^.*$'
          minLength: 1
          maxLength: 500
...